+++ tools/cabana/canmessages.cc
@@ -7,7 +7,6 @@
 
 Q_DECLARE_METATYPE(std::vector<CanData>);
 
-Settings settings;
 CANMessages *can = nullptr;
 
 CANMessages::CANMessages(QObject *parent) : QObject(parent) {
@@ -158,27 +157,4 @@ void CANMessages::resetRange() {
 
 void CANMessages::settingChanged() {
   replay->setSegmentCacheLimit(settings.cached_segment_limit);
-}
-
-// Settings
-
-Settings::Settings() {
-  load();
-}
-
-void Settings::save() {
-  QSettings s("settings", QSettings::IniFormat);
-  s.setValue("fps", fps);
-  s.setValue("log_size", can_msg_log_size);
-  s.setValue("cached_segment", cached_segment_limit);
-  s.setValue("chart_height", chart_height);
-  emit changed();
-}
-
-void Settings::load() {
-  QSettings s("settings", QSettings::IniFormat);
-  fps = s.value("fps", 10).toInt();
-  can_msg_log_size = s.value("log_size", 100).toInt();
-  cached_segment_limit = s.value("cached_segment", 3).toInt();
-  chart_height = s.value("chart_height", 200).toInt();
 }
